--- server/src-lib/Hasura/RQL/DDL/Schema/Cache.hs-boot
@@ -1,9 +1,0 @@
-module Hasura.RQL.DDL.Schema.Cache where
-
-import           Hasura.Prelude
-
-import           Hasura.Db
-import           Hasura.RQL.Types
-
-withMetadataCheck :: (MonadTx m, CacheRWM m) => Bool -> m a -> m a
-purgeDependentObject :: (MonadTx m) => SchemaObjId -> m ()
