+++ server/src-lib/Hasura/Server/Migrate.hs
@@ -32,6 +32,7 @@ import qualified Language.Haskell.TH.Syntax    as TH
 import           Hasura.Logging                (Hasura, LogLevel (..), ToEngineLog (..))
 import           Hasura.RQL.DDL.Schema
 import           Hasura.RQL.Types
+import           Hasura.RQL.Types.Run
 import           Hasura.Server.Logging         (StartupLog (..))
 import           Hasura.Server.Migrate.Version (latestCatalogVersion,
                                                 latestCatalogVersionString)
