+++ server/src-lib/Hasura/Server/Version.hs
@@ -1,4 +1,4 @@
--- {-# OPTIONS_GHC -fforce-recomp #-}
+{-# OPTIONS_GHC -fforce-recomp #-}
 module Hasura.Server.Version
   ( currentVersion
   , consoleVersion
